moved create user logic to UserService; added user on creation to the members group
