+++ src/domain/user/user.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { ProfileService } from '../profile/profile.service';
+import { UserInput } from './user.dto';
 import { User } from './user.entity';
 import { IUser } from './user.interface';
 
@@ -21,5 +22,22 @@ export class UserService {
 
   async getUserByEmail(email: string): Promise<IUser | undefined> {
     return User.findOne({ email: email });
+  }
+
+  async createUser(userData: UserInput): Promise<IUser> {
+    // Check if a user with this email already exists
+    const newUserEmail = userData.email;
+    const existingUser = await this.getUserByEmail(newUserEmail);
+
+    if (existingUser)
+      throw new Error(
+        `Already have a user with the provided email address: ${newUserEmail}`
+      );
+
+    // Ok to create a new user + save
+    const user = User.create(userData);
+    await user.save();
+
+    return user;
   }
 }
